% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlow.R
\name{plotFlow}
\alias{plotFlow}
\title{plotFlow}
\usage{
plotFlow(
  finalCells,
  dose.levels,
  plot.settings,
  cutoffs,
  xlab = "Annexin V - FITC",
  ylab = "DNA - DAPI"
)
}
\description{
Get Flow Cytometry plots
}
