% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFlowData.R
\name{readFlowData}
\alias{readFlowData}
\title{readFlowData}
\usage{
readFlowData(
  files = choose.files(filters = matrix(nrow = 1, data = c("FCS files", "*.fcs"))),
  dir,
  cutoffs = NULL,
  sample,
  date,
  plot.settings,
  analyzeCellCycle = F,
  time.name = "Time",
  channels = NULL
)
}
\description{
Read fcs files
}
