% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRFlowData.R
\name{readRFlowData}
\alias{readRFlowData}
\title{readRFlowData}
\usage{
readRFlowData(
  flowData,
  channels = c("B1-H", "V1-H"),
  dose.levels,
  time0 = 0,
  EDTA.point = NA,
  sample,
  date,
  type = "VIA",
  diff.cells = F
)
}
\description{
Read flow data from rData file for line plots
}
