% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCellCycle.R
\name{runCellCycle}
\alias{runCellCycle}
\title{reunCellCycle}
\usage{
runCellCycle(
  flowData,
  channels = "V1-H",
  dose.levels,
  sample,
  date,
  breaks = breaks,
  xlab = "DNA - DAPI",
  ylab = "BrdU - FITC",
  plot.settings,
  dir
)
}
\description{
Plot Cell Cycle Data
}
