% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFlow.R
\name{runFlow}
\alias{runFlow}
\title{runFlow}
\usage{
runFlow(
  dir,
  dose.levels,
  cutoffs = NULL,
  sample,
  date,
  plot.settings,
  ctrl,
  time0 = 0,
  analyzeCellCycle = F,
  time.name = "Time",
  channels = NULL
)
}
\description{
Main Flow Cytometry functions
}
