% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLinePlots.R
\name{runLinePlots}
\alias{runLinePlots}
\title{runLinePlots}
\usage{
runLinePlots(
  finalCells,
  plotDose = FALSE,
  dose.levels,
  channels = c("B1-H", "V1-H"),
  time0 = 0,
  sample,
  date,
  cutoffs,
  selectedTimepoints = NA,
  dir,
  breaks.via = c(2.5, 3.1),
  diff.cells = F
)
}
\description{
Main plotly line plot function
}
