% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFlowCells.R
\name{saveFlowCells}
\alias{saveFlowCells}
\title{saveFlowCells}
\usage{
saveFlowCells(Cells, flow.dir, dose.levels, ctrl, time0 = 0)
}
\description{
Add time zero to flow cells and return rData flow files
}
