% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeViabilityData.R
\name{summarizeViabilityData}
\alias{summarizeViabilityData}
\title{summarizeViabilityData}
\usage{
summarizeViabilityData(
  rawData,
  dose.levels,
  time.levels,
  channels = c("FL1-H", "FL18-H"),
  breaks.via = c(3, 3.2)
)
}
\description{
Summarize viability Data from dataframe
}
